Additional PCI password rules, and tests form them
